fix bugs of check recipe available

+++ brewDay/src/brewDay/RecipeIngredient.java
@@ -1,6 +1,6 @@
 package brewDay;
 public class RecipeIngredient extends Ingredient {
-	public RecipeIngredient(String name, float amount, String unit, int id) {
+	public RecipeIngredient(String name, double amount, String unit, int id) {
 		super(name, amount, unit,id);
 	}
